Rename AutoBindJaxRsWebServerModule > JaxRsWebServerModule
